Fix SAPI shutdown
This makes it so each instance of Embed holds an Arc so that the
underlying SAPI may be dropped when there are no more Embeds.
A Weak is held to attempt to upgrade an existing SAPI for any
future Embed constructions, but will construct a new SAPI any time
one is unavailable via the Weak.